requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
Django==4.2.7
django-bootstrap-v5==1.0.11
django-prometheus==2.3.1
prometheus-client==0.18.0
psycopg2-binary==2.9.9
soupsieve==2.5
sqlparse==0.4.4
oauthlib
requests
requests-oauthlib
qrcode
pyotp
djangorestframework-simplejwt
PyJWT